requirements.txt
streamlit 
flask
langchain
langchain-groq
python-dotenv